Программа для практической работы 3 попытка
