+++ src/main/java/Main.java
@@ -1,62 +1,4 @@
 import java.util.Scanner;
-import java.util.ArrayList;
-
-
-class Product {
-    float cost;
-    String name;
-
-    Product(float cost, String name) {
-        this.cost = cost;
-        this.name = name;
-    }
-}
-
-class Calculator {
-    float sum;
-    ArrayList<String> products = new ArrayList<>();
-
-    Calculator() {
-
-    }
-
-    public void addProduct(Product product) {
-        products.add( "\n" + product.name + " " + Float.toString(product.cost) + "\n");
-        sum += product.cost;
-    }
-
-    public void print() {
-        if (products.size() == 0) {
-            System.out.println("Дорогой пользователь, ты не ввёл ни одного товара!");
-            return;
-        }
-        System.out.println(products);
-    }
-}
-
-
-class Formatter{
-    public static String format(double sum, int  amount){
-        int numb= (int) ((sum/amount)%100);
-        if(numb>=11&&numb<=19){
-            return "рублей";
-        }
-        int digit= (int) ((sum/amount)%10);
-        switch (digit){
-            case 1:
-                return "рубль";
-            case 2:
-            case 3:
-            case 4:
-                return "рубля";
-            default:
-                return "рублей";
-        }
-    }
-}
-
-
-
 class Main {
 
     public static void printMenu() {
@@ -77,8 +19,8 @@ class Main {
                     System.out.println("Корректно вводите данные! Кол-во людей должно быть больше 1!");
                 }
             } else {
-                System.out.println("Извините, но это явно не число. Перезапустите программу и попробуйте снова!");
-                return;
+                System.out.println("Извините, но это явно не число. Попробуйте снова!");
+                String trash = scanner.next();
             }
         }
         while (amount <= 1);
@@ -103,16 +45,24 @@ class Main {
                         System.out.println("Дорогой пользователь, введи корректную цену товара)");
                     }
                     break;
+                
                 case "2":
                     calc.print();
                     float numb =  (calc.sum/amount);
                     String formated = String.format("%.2f",numb);
                     System.out.println("Пока каждый из вашей компании должен будет за товары " + formated+ " "  + Formatter.format(calc.sum, amount));
                     break;
+
                 case "завершить":
+                    calc.print();
+                    float numer =  (calc.sum/amount);
+                    String format = String.format("%.2f",numer);
+                    System.out.println("Пока каждый из вашей компании должен будет за товары " + format+ " "  + Formatter.format(calc.sum, amount));
                     scanner.close();
                     return;
 
+                default:
+                    throw new IllegalStateException("Unexpected value: " + num);
             }
         }
     }
