+++ src/main/java/Product.java
@@ -1,0 +1,9 @@
+class Product {
+    float cost;
+    String name;
+
+    Product(float cost, String name) {
+        this.cost = cost;
+        this.name = name;
+    }
+}
